further work on the ui

+++ app/src/main/java/ie/noel/dunsceal/persistence/db/DunStoreRoom.kt
@@ -4,6 +4,7 @@ import android.content.Context
 import androidx.room.Room
 import ie.noel.dunsceal.models.Dun
 import ie.noel.dunsceal.models.DunStore
+import ie.noel.dunsceal.models.entity.Dun
 import ie.noel.dunsceal.persistence.db.dao.DunDao
 
 
@@ -22,7 +23,7 @@ class DunStoreRoom(val context: Context) : DunStore {
     return dao.findAll()
   }
 
-  override fun findById(id: Int): Dun? {
+  override fun findById(id: Long): Dun? {
     return dao.findById(id)
   }
